refactored your code to follow Oppia's proven pattern
