+++ scripts/check_for_unresolved_todos.py
@@ -16,11 +16,269 @@
 from __future__ import annotations
 
 import argparse
+import json
 import os
+import subprocess
+import textwrap
+import urllib.request
 
-from scripts import github_api, todo_finder
+from scripts import todo_finder
 
-from typing import List, Optional
+from typing import Any, Dict, List, Optional, TypedDict
+
+# GitHub repository information.
+REPOSITORY_OWNER = os.environ.get('GITHUB_REPOSITORY', 'oppia/oppia').split(
+    '/'
+)[0]
+REPOSITORY_NAME = os.environ.get('GITHUB_REPOSITORY', 'oppia/oppia').split('/')[
+    1
+]
+GITHUB_API_URL = 'https://api.github.com'
+
+
+class GitHubIssueDict(TypedDict):
+    """Dict representation of a GitHub issue."""
+
+    body: str
+    number: int
+    title: str
+
+
+class GitHubCommentDict(TypedDict):
+    """Dict representation of a GitHub comment."""
+
+    body: str
+
+
+def deep_get(data: Optional[Dict[str, Any]], keys: List[str]) -> Any:
+    """Gets a value from a nested dictionary. If the key is not found, it
+    returns None.
+
+    Args:
+        data: dict. The dictionary to get the value from.
+        keys: List[str]. The keys to traverse the dictionary.
+
+    Returns:
+        Any. The value from the nested dictionary.
+    """
+    if not keys or data is None:
+        return data
+    return deep_get(data.get(keys[0]), keys[1:])
+
+
+def get_github_auth_token() -> str:
+    """Gets the GitHub auth token.
+
+    Returns:
+        str. The GitHub auth token.
+
+    Raises:
+        RuntimeError. GitHub CLI is not installed.
+        RuntimeError. Failed to get GitHub auth token.
+    """
+    env = os.environ.copy()
+    # Prefer environment variables (used by GitHub Actions + can be set locally)
+    # and fall back to GitHub CLI token resolution.
+    github_token = env.get('GH_TOKEN') or env.get('GITHUB_TOKEN')
+    if github_token:
+        return github_token
+
+    process = subprocess.run(
+        ['gh', 'help'],
+        env=env,
+        check=False,
+        encoding='utf-8',
+        stdout=subprocess.DEVNULL,
+        stderr=subprocess.STDOUT,
+    )
+    if process.returncode:
+        raise RuntimeError(
+            'GitHub CLI is not installed. Please install the GitHub CLI '
+            'before running GitHub API functions.'
+        )
+    process = subprocess.run(
+        ['gh', 'auth', 'token'],
+        capture_output=True,
+        encoding='utf-8',
+        env=env,
+        check=False,
+    )
+    if process.returncode:
+        raise RuntimeError(
+            'Failed to get GitHub auth token from the GitHub CLI.'
+        )
+    return process.stdout.strip()
+
+
+def get_github_api_authorization_header() -> str:
+    """Formats the GitHub auth token to be used as an authorization header."""
+    return f'Bearer {get_github_auth_token()}'
+
+
+def run_graphql_query(query: str) -> Dict[str, Any]:
+    """Runs a GraphQL query on the Oppia repository.
+
+    Args:
+        query: str. The query to run.
+
+    Returns:
+        Dict[str, Any]. The response from the GraphQL query in JSON.
+
+    Raises:
+        RuntimeError. Failed to run GraphQL query due to a request error.
+        RuntimeError. Failed to run GraphQL query due to an API error.
+    """
+    constructed_query = (
+        textwrap.dedent(
+            """
+        query {
+            repository(owner: "%s", name: "%s") {
+                %s
+            }
+        }
+        """
+        )
+        % (REPOSITORY_OWNER, REPOSITORY_NAME, query)
+    )
+    try:
+        url = f'{GITHUB_API_URL}/graphql'
+        headers = {
+            'Authorization': get_github_api_authorization_header(),
+            'Content-Type': 'application/json',
+        }
+        request_data = json.dumps({'query': constructed_query}).encode('utf-8')
+        request = urllib.request.Request(url, request_data, headers)
+        with urllib.request.urlopen(request) as response:
+            if response.getcode() == 200:
+                data: Dict[str, Any] = json.loads(
+                    response.read().decode('utf-8')
+                ).get('data')
+                return data
+            else:
+                raise RuntimeError(
+                    'Failed to run the GraphQL query due to an '
+                    + 'API error: %s' % response.read().decode('utf-8')
+                )
+    except Exception as e:
+        raise RuntimeError(
+            'Failed to run the GraphQL query due to a '
+            + 'request error: %s' % e
+        ) from e
+
+
+def fetch_linked_issues_for_pull_request(
+    pull_request: int,
+) -> List[GitHubIssueDict]:
+    """Fetches the linked issues for a pull request.
+
+    Args:
+        pull_request: int. The pull request number.
+
+    Returns:
+        List[GitHubIssueDict]. The linked issues for the pull request.
+
+    Raises:
+        Exception. The pull request does not exist.
+    """
+    query = (
+        textwrap.dedent(
+            """
+        pullRequest(number: %s) {
+            closingIssuesReferences(first: 50) {
+                nodes {
+                    body
+                    number
+                    title
+                }
+            }
+        }
+        """
+        )
+        % pull_request
+    )
+    response = run_graphql_query(query)
+    linked_issues: List[GitHubIssueDict] = deep_get(
+        response,
+        ['repository', 'pullRequest', 'closingIssuesReferences', 'nodes'],
+    )
+    if linked_issues is None:
+        raise Exception('The pull request (#%s) does not exist.' % pull_request)
+    return linked_issues
+
+
+def fetch_latest_comment_for_issue(
+    issue_number: int,
+) -> Optional[GitHubCommentDict]:
+    """Fetches the latest comment from an issue.
+
+    Args:
+        issue_number: int. The issue number.
+
+    Returns:
+        Optional[GitHubCommentDict]. The latest comment from the issue.
+
+    Raises:
+        Exception. The issue does not exist.
+    """
+    query = (
+        textwrap.dedent(
+            """
+        issue(number: %s) {
+            comments(last: 1) {
+                nodes {
+                    body
+                }
+            }
+        }
+        """
+        )
+        % issue_number
+    )
+    response = run_graphql_query(query)
+    latest_comment_nodes = deep_get(
+        response, ['repository', 'issue', 'comments', 'nodes']
+    )
+    if latest_comment_nodes is None:
+        raise Exception('The issue (#%s) does not exist.' % issue_number)
+    return next(iter(latest_comment_nodes), None)
+
+
+def fetch_latest_comment_from_pull_request(
+    pull_request: int,
+) -> Optional[GitHubCommentDict]:
+    """Fetches the latest comment from a pull request.
+
+    Args:
+        pull_request: int. The pull request number.
+
+    Returns:
+        Optional[GitHubCommentDict]. The latest comment from the pull request.
+
+    Raises:
+        Exception. The pull request does not exist.
+    """
+    query = (
+        textwrap.dedent(
+            """
+        pullRequest(number: %s) {
+            comments(last: 1) {
+                nodes {
+                    body
+                }
+            }
+        }
+        """
+        )
+        % pull_request
+    )
+    response = run_graphql_query(query)
+    latest_comment_nodes = deep_get(
+        response, ['repository', 'pullRequest', 'comments', 'nodes']
+    )
+    if latest_comment_nodes is None:
+        raise Exception('The pull request (#%s) does not exist.' % pull_request)
+    return next(iter(latest_comment_nodes), None)
+
 
 # TODO(#6): Temporary workflow test first.
 # TODO(#6): Temporary workflow test SECOND.
@@ -152,7 +410,7 @@ def main(args: Optional[List[str]] = None) -> None:
     if parsed_args.issue:
         issues_to_check.append(parsed_args.issue)
     if parsed_args.pull_request:
-        linked_issues = github_api.fetch_linked_issues_for_pull_request(
+        linked_issues = fetch_linked_issues_for_pull_request(
             parsed_args.pull_request
         )
         for linked_issue in linked_issues:
